+++ wsei-todo-app-master/src/components/BottomNavigation.tsx
@@ -36,7 +36,7 @@ function BottomTabs(){
                 options={{
                     tabBarLabel: 'TodoList',
                     tabBarIcon: ({color, size}) => (
-                        <MaterialCommunityIcons name = "TodoList" color={color} size = {size} />
+                        <MaterialCommunityIcons name = "plus-circle" color={color} size = {size} />
                     ),
                 }}
             />